requirements.txt

pyyaml
torch>=2.0
torchvision
pytorchvideo
numpy
opencv-python
scikit-learn
tqdm
einops
